% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{altOfSun}
\alias{altOfSun}
\title{Find sun heigth and sun rise. Thanks to Espen Johnsen who provided the code.}
\usage{
altOfSun(min, hour, day, month, lat, lon)
}
\description{
Find sun heigth and sun rise. Thanks to Espen Johnsen who provided the code.
}
