% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{createMesh}
\alias{createMesh}
\title{Create mesh}
\usage{
createMesh(cutoff = NULL, cbound = NULL)
}
\arguments{
\item{cutoff}{Parameter providing density of mesh}

\item{cbound}{Boundary parameter}
}
\value{
Mesh used
}
\description{
This function creates the mesh to represent the Matern covariance strucutre
}
