% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataInYear}
\alias{dataInYear}
\title{dataInYear}
\usage{
dataInYear(year, conf, data)
}
\arguments{
\item{year}{Year which is setup}

\item{conf}{Configurations}

\item{data}{Data from all years}
}
\value{
data-object
}
\description{
dataInYear
}
\details{
Prepares data from Rstox in a single year obtained by getBaseLine for the modeling routine.
}
