% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{jit}
\alias{jit}
\title{jit}
\usage{
jit(run, njit, ncores = 1, sd = 0.2)
}
\arguments{
\item{run}{The result of running fitModel}

\item{njit}{Number of jitter runs}

\item{ncores}{Number of cores to use}

\item{sd}{Standard deviation of noise to start values}
}
\value{

}
\description{
jit
}
