% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCPUE.R
\name{mCPUESimultanious}
\alias{mCPUESimultanious}
\title{mCPUESimultanious}
\usage{
mCPUESimultanious(run, rep = NULL, biasCorrect = FALSE)
}
\arguments{
\item{run}{Estimated model}
}
\value{


Estimated index in all stratas simulaniously, and in selected areas provided to the C-side
}
\description{
mCPUESimultanious
}
