% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotResults}
\alias{plotResults}
\title{plotResults Plot results}
\usage{
plotResults(run, what = NULL, range = c(-10, 6), legend = TRUE)
}
\arguments{
\item{run}{fitted object returned by \code{\link{fitModel}}}

\item{what}{what to plot, options in first element "sunAlt", "depth", "S", "ST" and "SST". If spatial effect is plotted, year and length must be provided}
}
\description{
plotResults Plot results
}
