% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.stim}
\alias{print.stim}
\title{Print stim object}
\usage{
\method{print}{stim}(x, ...)
}
\arguments{
\item{x}{}
}
\description{
Print stim object
}
\details{
Print log-likelihood and the main convergence criteria
}
