% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runwithout}
\alias{runwithout}
\title{runwithout Fit stim and leave out observation indices as provided in fold}
\usage{
runwithout(data, par, conf, map, fold)
}
\arguments{
\item{data}{Data}

\item{par}{Parameters}

\item{conf}{Configurations}

\item{map}{map-argument}

\item{fold}{Vector of data indices to remove when fitting stim}
}
\description{
runwithout Fit stim and leave out observation indices as provided in fold
}
