% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{setPar}
\alias{setPar}
\title{Initialie values for all model parameters and random effects.}
\usage{
setPar(data, conf)
}
\arguments{
\item{data}{Data}

\item{conf}{Configurations}
}
\value{
a list containing initial values for all model parameters and random effects in the model.
}
\description{
Initialie values for all model parameters and random effects.
}
