% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate.stim}
\alias{simulate.stim}
\title{Simulate from a STIM object}
\usage{
\method{simulate}{stim}(run, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{run}{}

\item{nsim}{}

\item{seed}{}
}
\description{
Simulate from a STIM object
}
\details{
Simulate observations from STIM as explained in Breiviek et al. (2020).
}
